Move `IPoll` definition to interfaces.ts.
